Restructured the project
