+++ src/main/java/maas/agents/BookBuyerAgent.java
@@ -1,4 +1,4 @@
-package maas;
+package maas.agents;
 
 import jade.content.lang.Codec;
 import jade.content.lang.sl.SLCodec;
