+++ src/main/java/maas/utils/OrderCustomerIdComparator.java
@@ -1,6 +1,8 @@
-package maas;
+package maas.utils;
 
 import java.util.Comparator;
+
+import maas.objects.Order;
 
 public class OrderCustomerIdComparator implements Comparator<Order>{
